Implemented "formatter" framework for formatting output for various statusbars and implemented a formatter for LemonBar
To run gostatus for a specific statusbar, you implement a formatter for it, and pass that into the segments

A formatter is an immutable structure that allows you to wrap output in
statusbar-specific formatting tags. Formatting operations can be chained
together to build complex combinations of formatting tags. This makes it
easier to apply a universal "style" to each segment.